tutorial-generator: all-in-one generator
A Python script at `tutorial-generator/gen.py` takes an input
directory as an argument. It looks for all `*-instructions.txt`
files, where the prefix is a number to determine step order.

For each step, an optional `#-prepare.sh` script can be present
to stage things (e.g. change a source file) and a `#-check.sh`
script will run after outputting instructions to provide feedback
to the user about status. To make this easier, there is an
`api-check.sh` script that can be invoked from user-provided
scripts which handles calling Tilt API and blocking until a
condition is met for a particular resource.

The `gen.py` script bundles all of this INTO a `workshop.tiltfile`
in the working directory where it's called. This Tiltfile can
then be included into a parent Tiltfile and will no-op unless
the `WORKSHOP` env var is set. Once active, it creates a meta
"workshop" resource that bootstraps the tutorial.